+++ src/Core/Algorithms/Examples/SortMatrixAlgo.h
@@ -3,10 +3,10 @@
 
    The MIT License
 
-   Copyright (c) 2009 Scientific Computing and Imaging Institute,
+   Copyright (c) 2015 Scientific Computing and Imaging Institute,
    University of Utah.
 
-   
+   License for the specific language governing rights and limitations under
    Permission is hereby granted, free of charge, to any person obtaining a
    copy of this software and associated documentation files (the "Software"),
    to deal in the Software without restriction, including without limitation
@@ -29,37 +29,39 @@
 #ifndef CORE_ALGORITHMS_MATH_SortMatrixALGO_H
 #define CORE_ALGORITHMS_MATH_SortMatrixALGO_H
 
-#include <Core/Datatypes/Matrix.h>
-#include <Core/Datatypes/DenseMatrix.h>
-#include <Core/Datatypes/DenseColumnMatrix.h>
+#include<Core/Datatypes/Matrix.h>
+#include<Core/Datatypes/DenseMatrix.h>
+#include<Core/Datatypes/DenseColumnMatrix.h>
 
+#include<string>
+#include<sstream>
+#include<vector>
+#include<algorithm>
 
-#include <string>
-#include <sstream>
-#include <vector>
-#include <algorithm>
+#include<Core/Algorithms/Base/AlgorithmVariableNames.h>
+#include<Core/Algorithms/Base/AlgorithmBase.h>
+#include<Core/Algorithms/Math/share.h>
 
-#include <Core/Algorithms/Base/AlgorithmVariableNames.h>
-#include <Core/Algorithms/Base/AlgorithmBase.h>
-#include <Core/Algorithms/Math/share.h>
+namespace SCIRun{
+    namespace Core{
+        namespace Algorithms{
+            namespace Math{
+                class SCISHARE SortMatrixAlgo : public AlgorithmBase
+                {
+                public:
+                    SortMatrixAlgo();
+                    
+                    AlgorithmOutput run(const AlgorithmInput& input) const;
+                    
+                    bool Sort(Datatypes::DenseMatrixHandle input, Datatypes::DenseMatrixHandle& output, int method) const;
+                    
+                    bool Quicksort(double* input, index_type lo, index_type hi) const;
+                    
+                    index_type Partition(double* input, index_type lo, index_type hi) const;
+                };
+            }
+        }
+    }
+}
 
-namespace SCIRun {
-namespace Core {
-namespace Algorithms {
-namespace Math {
-
-class SCISHARE SortMatrixAlgo : public AlgorithmBase
-{
-
-  public:
-    SortMatrixAlgo();
-    AlgorithmOutput run(const AlgorithmInput& input) const;
-  
-    bool Sort(Datatypes::DenseMatrixHandle input, Datatypes::DenseMatrixHandle& output, int method) const;
-  
-    bool Quicksort(double* input, index_type lo, index_type hi) const;
-    index_type Partition(double* input, index_type lo, index_type hi) const;
-};
-
-      }}}}// end SCIRun namespace
 #endif
