--- src/Core/Algorithms/Math/SortMatrixAlgo.h
@@ -1,67 +1,0 @@
-/*
-   For more information, please see: http://software.sci.utah.edu
-
-   The MIT License
-
-   Copyright (c) 2015 Scientific Computing and Imaging Institute,
-   University of Utah.
-
-   License for the specific language governing rights and limitations under
-   Permission is hereby granted, free of charge, to any person obtaining a
-   copy of this software and associated documentation files (the "Software"),
-   to deal in the Software without restriction, including without limitation
-   the rights to use, copy, modify, merge, publish, distribute, sublicense,
-   and/or sell copies of the Software, and to permit persons to whom the
-   Software is furnished to do so, subject to the following conditions:
-
-   The above copyright notice and this permission notice shall be included
-   in all copies or substantial portions of the Software.
-
-   THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS
-   OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-   FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL
-   THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-   LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING
-   FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER
-   DEALINGS IN THE SOFTWARE.
-*/
-
-#ifndef CORE_ALGORITHMS_MATH_SortMatrixALGO_H
-#define CORE_ALGORITHMS_MATH_SortMatrixALGO_H
-
-#include<Core/Datatypes/Matrix.h>
-#include<Core/Datatypes/DenseMatrix.h>
-#include<Core/Datatypes/DenseColumnMatrix.h>
-
-#include<string>
-#include<sstream>
-#include<vector>
-#include<algorithm>
-
-#include<Core/Algorithms/Base/AlgorithmVariableNames.h>
-#include<Core/Algorithms/Base/AlgorithmBase.h>
-#include<Core/Algorithms/Math/share.h>
-
-namespace SCIRun{
-    namespace Core{
-        namespace Algorithms{
-            namespace Math{
-                class SCISHARE SortMatrixAlgo : public AlgorithmBase
-                {
-                public:
-                    SortMatrixAlgo();
-                    
-                    AlgorithmOutput run(const AlgorithmInput& input) const;
-                    
-                    bool Sort(Datatypes::DenseMatrixHandle input, Datatypes::DenseMatrixHandle& output, int method) const;
-                    
-                    bool Quicksort(double* input, index_type lo, index_type hi) const;
-                    
-                    index_type Partition(double* input, index_type lo, index_type hi) const;
-                };
-            }
-        }
-    }
-}
-
-#endif
