+++ src/Interface/Modules/Examples/TestModuleSimpleUIDialog.ui
@@ -35,15 +35,15 @@
       <property name="fieldGrowthPolicy">
        <enum>QFormLayout::FieldsStayAtSizeHint</enum>
       </property>
-      <item row="0" column="0">
+      <item row="1" column="1">
+       <widget class="QLineEdit" name="inputstring_"/>
+      </item>
+      <item row="1" column="0">
        <widget class="QLabel" name="label_3">
         <property name="text">
-         <string>My String:</string>
+         <string>String Entered</string>
         </property>
        </widget>
-      </item>
-      <item row="0" column="1">
-       <widget class="QLineEdit" name="inputstring_"/>
       </item>
      </layout>
     </widget>
