+++ src/Interface/Modules/Math/CMakeLists.txt
@@ -50,7 +50,6 @@ SET(Interface_Modules_Math_FORMS
   DisplayHistogram.ui
   BasicPlotter.ui
   AdvancedPlotter.ui
-  SortMatrixDialog.ui
   GenerateStandardMatrixDialog.ui
 )
 
@@ -80,7 +79,6 @@ SET(Interface_Modules_Math_HEADERS
   BasicPlotterDialog.h
   AdvancedPlotterDialog.h
   PlotDialog.h
-  SortMatrixDialog.h
   GenerateStandardMatrixDialog.h
 )
 
@@ -109,7 +107,6 @@ SET(Interface_Modules_Math_SOURCES
   BasicPlotterDialog.cc
   AdvancedPlotterDialog.cc
   PlotDialog.cc
-  SortMatrixDialog.cc
   GenerateStandardMatrixDialog.cc
 )
