+++ src/Interface/Modules/String/CMakeLists.txt
@@ -34,7 +34,6 @@ SET(Interface_Modules_String_FORMS
   PrintMatrixIntoStringDialog.ui
   PrintStringIntoStringDialog.ui
   ReportStringInfo.ui
-  TestModuleSimpleUIDialog.ui
 )
 
 SET(Interface_Modules_String_HEADERS
@@ -45,7 +44,6 @@ SET(Interface_Modules_String_HEADERS
   PrintMatrixIntoStringDialog.h
   PrintStringIntoStringDialog.h
   ReportStringInfoDialog.h
-  TestModuleSimpleUIDialog.h
 )
 
 SET(Interface_Modules_String_SOURCES
@@ -55,7 +53,6 @@ SET(Interface_Modules_String_SOURCES
   PrintMatrixIntoStringDialog.cc
   PrintStringIntoStringDialog.cc
   ReportStringInfoDialog.cc
-  TestModuleSimpleUIDialog.cc
 )
 
 QT_WRAP_UI(Interface_Modules_String_FORMS_HEADERS "${Interface_Modules_String_FORMS}")
