+++ src/Modules/Math/CMakeLists.txt
@@ -51,7 +51,6 @@ SET(Modules_Math_SRCS
   DisplayHistogram.cc
   BasicPlotter.cc
   AdvancedPlotter.cc
-  SortMatrix.cc
   GenerateStandardMatrix.cc
 )
 
@@ -81,7 +80,6 @@ SET(Modules_Math_HEADERS
   DisplayHistogram.h
   BasicPlotter.h
   AdvancedPlotter.h
-  SortMatrix.h
   GenerateStandardMatrix.h
 )
